version 0.2
Updated to fix issue that menu bar, toolbar and status_bar change size if you change window size.
  Fixed with property vexpand to FALSE.
Now the account buttons are inside a GtkScrolledWindow (that needs a child GtkViewport to add a container),
  so you can resize the window and scroll to view any button.
Now I use GtkBox in both containers because is not necessary a GtkGrid.
Increased maximum number of accounts to 1000.
Removed gtk_widget_show for each widget because the window has gtk_widget_show_all (window).